+++ src/components/index.ts
@@ -1,1 +1,3 @@
-export * from "./flow";
+export * from "./FlowPanel";
+export * from "./ImageStore";
+export * from "./dialog";
